pyproject.toml
[tool.poetry]
name = "ssl-cert-gen"
version = "0.1.0"
description = "Generate X509 Certificates for Dev"
authors = ["CTG"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.10.2"
pyopenssl = "^22.1.0"
PyYAML = "^6.0"


[tool.poetry.dev-dependencies]
pylint = "^2.15.9"
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
